Crate walrus_meter

walrus-meter

This crate provides instrumentation-based metering for Wasm modules using the walrus library.

Provide the name of a function to call, and the function will be called immediately before executing a sequence of instructions with that sequence’s cost.

Modules

costs

Various prebuilt cost strategies.

Traits

Costs

A trait that encodes the costs of different Wasm operations.

Functions

instrument

Injects the named spend function and instrument all functions to call it according to the given costs.

Type Aliases

Error

The type of errors raised by this crate.